# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Login Backend API
  version: 0.1.0
  description: Simple authentication API with JWT access & refresh tokens.
servers:
  - url: http://localhost:3000
paths:
...
  /auth/refresh:
    post:
      summary: Refresh tokens
      description: Uses refresh_token httpOnly cookie.
      responses:
        '200':
          description: OK - new tokens issued
          headers:
            Set-Cookie:
              schema:
                type: string
              description: New access_token & refresh_token cookies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Invalid or missing refresh token
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
    UserResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'